+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * EXOMailboxCalendarFolder
   * Add support for non-English calendar folder names during export
     FIXES [#4056](https://github.com/microsoft/Microsoft365DSC/issues/4056)
+* EXOTransportRule
+  * Fix export of enabled state
+    FIXES [#3932](https://github.com/microsoft/Microsoft365DSC/issues/3932)
 * IntuneDeviceConfigurationCustomPolicyWindows10
   * Fix issue deploying decrypted OmaSettings to another tenant
     FIXES [#4083](https://github.com/microsoft/Microsoft365DSC/issues/4083)